Relocate functions in template package
